Add Sdk version by default to all telemetry. Remove SdkVersionTelemetryInitializer so it cannot be removed.
